requirements.txt
alembic==1.17.2
fastapi==0.124.4
pydantic[email]==2.12.2
pandas==2.3.3
python-dotenv==1.2.1
python-multipart==0.0.9
psycopg2-binary==2.9.11
SQLAlchemy==2.0.45
uvicorn==0.28.0
